requirements.txt
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.2
newspaper3k==0.2.8
nltk==3.8.1
textblob==0.17.1
transformers==4.36.2
torch==2.1.2
python-dotenv==1.0.0
feedparser==6.0.10
lxml==4.9.3
fake-useragent==1.4.0
tqdm==4.66.1
matplotlib==3.7.2
seaborn==0.13.0
plotly==5.17.0
dash==2.14.2
apscheduler==3.10.4
yfinance==0.2.28
googlesearch-python==1.2.3
